refactor(TimePicker): migrate JSX to TypeScript
